added workdir option to ktest and a test
-w can be used to specify a directory that can be written to.
   Useful if the kernel directory is not writable or when
   running multiple tests in parallel

Also added a test that checks that the basic ktest operations
work. It boots a vm, which sleeps and waits for an ssh connection
to kill it. The test passes if the sleep actually does get killed.

As this test requires a kernel ktest now requires linux-bcache
to be built first, rather than the other way around. As a result
the kernel tests are now driven out of this package (currently
only tests/bcache/bcache.ktest).

It also makes more sense to run the tests within the package
that defines them, rather than in a separate package. Previously
one had to edit two packages to get a test to run, while now
tests are fully contained in the ktest package.

Note: adding .ktest as a suffix simplifies the makefile considerably.