Add a connect x3 virtual function to a vm
Use "vmstart genconx3" to create a database that has all the virtual functions
detached of a Mellanox connect x3 card.

Now you can start the vm with the "vmstart -conx3" to assign a vf to a
virtual machine.

Note the ip assignment is still not in scripts. That has to be done manually
as of now for the ib0 interface

Signed-off-by: Surbhi Palande <sap@daterainc.com>