ktest: fixes for timeout and interactive mode
The ktest script did not ensure that the check_timeout process was
killed when the user sent a signal to kill the ktest process.
As a result, interrupting a ktest run and starting a new could
result in the new invocation being prematurely killed by the lingering
timeout.

Also fix interactive mode. Previously if -I was passed, we would do
two things:
- disable parsing output for 'TEST SUCCESS'
- add code to /etc/rc to spawn a shell and then power off when the
  shell exits

The second part was broken however since the output of /etc/rc is
piped through a process that prefixes each line with a string, and
there's no tty. So the shell just exits and we power off immediately,
even if the test failed. This is not intended behavior.

With this patch, -I now behaves as follows:
- disable parsing output for 'TEST SUCCESS'
- disable test timeout, if one is set

Change-Id: Ie9531c094e9357621d0ad228251a4db017111fcf