ktest: Use new GCOV_PROFILE_<dir> feature for -c flag
The -c flag for enabling coverage now takes a directory name.
Multiple -c flags may be passed in to enable coverage on
multiple directories.

Also, fix 'usage' printout -- the function was defined too late
in the file, so running 'ktest' with no arguments would fail
with an error instead of printing usage.

Change-Id: Id276295cabd13dc0f0b54043fe7f004362ca71af