ktest: fix recent reboot test breakage and make tests look the same
This is a first step toward combining some of the reboot and non-reboot
workloads, but we have to make reboot tests reliable first.

In particular, the do_reboot call runs copy_to_host which kills the
fio instance. It is very important to handle the failed exit code
from fio correctly. On the last reboot however, we want to run fio
to completion and fail the test if it doesn't finish.

Also SIZE=large makes the random write phase of test_fio too long.
The SIZE stuff will be redone later, but make it small for now.

Change-Id: Idd69df849adc2b44a3e1df2a0485a45e2e5a2615