ktest: crash on timeout
- The top-level timeout in ktest adds an extra 20 seconds
- Internal timeout inside the VM writes c > /proc/sysrq-trigger
- Change ktest test to expect exit code 1 for both timeouts and failures

This will catch some, but not all, timeouts. If the VM is so wedged
that the internal timeout doesn't work, we won't get a deadlock, but
those are more rare.

If we encounter the latter case we can try to add a 'pre-timeout' to
ktest itself, but after trying to get that working and giving up due to
issues with job control and shared FDs and such, I decided the simpler
approach is better.

Change-Id: Ie3f97464f44cc664959f3fdbcbb12fedcae9d9b6