Revert change to make -j
build-test-kernel used to check the number of CPUS and pass that as the
make -j argument. We dropped that and started just calling make -j
because make got support for limiting based on load average, but it
appears that was a bad idea - this patch brings back the limit based on
nr of cpus.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>