root_image: support chroot not contained in /usr/sbin, /usr/bin, /sbin or /bin
Currently, `_chroot` performs "a restriction" of the `PATH` environment variable.

This is done to avoid mixing too much the host binaries inside the chroot, this makes sense.
The issue is that we need to load `chroot` from the host, therefore, we use `/usr/bin/env` to load them.

The problem is that, on NixOS, `/usr/bin/env chroot` will perform a PATH lookup and fail, our `/usr/bin/env`
is not "hardcoded" to the "current shell" environment no matter what `PATH` we do set and rightfully so.

The easy solution I suppose is to compute `CHROOT` before we restrict and reuse it as needed.