Introduce partial cross-compile of bcachefs-tools
When running ktest, the qemu vm needs to cross-compile bcachefs-tools.
this can be pretty slow, especially on emulated architectures.
to work around this, use the host compiler in advance.
This pre-compiles the C files into object files,
which will then be re-used on the target VM.

This feature needs fuse2fs FUSE binary to mount the ktest root image:
mounting non-root is not possible, but headers need to be included.

If the requirements aren't met, the VM needs to do everything.

Signed-off-by: jpsollie <janpieter.sollie@edpnet.be>