tests/puzzlefs: add a ktest for PuzzleFS
This test builds a PuzzleFS image using the PuzzleFS user space tools,
builds a Linux kernel with PuzzleFS enabled, mounts the PuzzleFS image
using the kernel driver and then compares the two filesystems: the
source filesystem used when building the PuzzleFS image and the final
PuzzleFS filesystem mounted directly from the PuzzleFS image.

This test is similar to the build_from_ext4 test in single_device.ktest.

Two packages are added to the root_image: capnproto required for
building the PuzzleFS user space tools and jq, which is used for
extracting the PuzzleFS image manifest from index.json.

Signed-off-by: Ariel Miculas <amiculas@cisco.com>
Signed-off-by: Kent Overstreet <kent.overstreet@linux.dev>